test: reorganize OpenAPI test structure
- Remove redundant old test files that were duplicating functionality
- Add simple_test.go with focused unit tests
- Consolidate test coverage while maintaining comprehensive integration tests
